fix(timeutils): fix darwin and windows cross-compilation
Use int32 Usec on Darwin and call SetSystemTime via kernel32 so static
release builds succeed for macOS arm64 and Windows from Linux.

+++ timeutils/timeutils_darwin.go
@@ -13,7 +13,7 @@ func SetSystemTime(t time.Time) error {
 	utc := t.UTC()
 	tv := unix.Timeval{
 		Sec:  utc.Unix(),
-		Usec: int64(utc.Nanosecond() / 1000),
+		Usec: int32(utc.Nanosecond() / 1000),
 	}
 	return unix.Settimeofday(&tv)
 }
